requirements.txt
PyQt6==6.6.1
websockets==12.0
pyusb==1.2.1
opencv-python==4.9.0.80
numpy==1.26.3
pillow==10.2.0
python-dotenv==1.0.1 